Rest of features, 100% working machine

+++ code/ArduinoAnalogPin.cpp
@@ -1,6 +1,6 @@
 #include "ArduinoAnalogPin.h"
 
-int ArduinoAnalogPin::read()
+short ArduinoAnalogPin::read()
 {
     return analogRead(getAddress());
 }
